requirements-simple.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
python-dotenv
pydantic
